# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PAD2 Lazuardy API
  version: 1.0.0
  description: |
    Dokumentasi API untuk backend Laravel PAD2 Lazuardy.

    Semua response mengikuti format standar:

    {
      "status": "success" | "error",
      "message": "string",
      "data": object|null,
      "errors": object|null
    }
servers:
  - url: https://api.example.com
...
paths:
  /login:
    post:
      tags: [Auth]
      summary: Login
      description: Login menggunakan email dan password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponseData'
        '401':
          description: Kredensial tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /logout:
    post:
      tags: [Auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me:
    get:
      tags: [Auth]
      summary: Get current user (role-based)
      description: |
        Response `data` bergantung pada role:

        - role=student -> `data.student`
        - role=tutor -> `data.tutor`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil user saat ini
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/MeStudentData'
                          - $ref: '#/components/schemas/MeTutorData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    patch:
      tags: [Auth]
      summary: Update current user (partial)
      description: Partial update. Kirim hanya field yang ingin diubah.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /student/review:
    get:
      tags: [Review]
      summary: List student reviews
      description: Hanya bisa diakses oleh student.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar review student
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [reviews]
                        properties:
                          reviews:
                            type: array
                            items:
                              $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags: [Review]
      summary: Create review (1 session = 1 review)
      description: |
        Rules:
        - rate: number (1-5)
        - 1 session = 1 review
        - hanya student yang boleh akses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsertRequest'
      responses:
        '201':
          description: Review tersimpan
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [review]
                        properties:
                          review:
                            $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'

    patch:
      tags: [Review]
      summary: Update review (upsert)
      description: Patch review by session+tutor context (implementation-specific).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsertRequest'
      responses:
        '200':
          description: Review tersimpan
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [review]
                        properties:
                          review:
                            $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /student/review/{tutorId}:
    get:
      tags: [Review]
      summary: Review detail by tutor
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: path
          required: true
          schema:
            type: integer
          example: 12
      responses:
        '200':
          description: Review tutor oleh student
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [review]
                        properties:
                          review:
                            $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                status: error
                message: Unauthorized
                data: null
                errors: null
    ForbiddenError:
      description: Forbidden (role-based access)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                status: error
                message: Forbidden
                data: null
                errors: null
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              value:
                status: error
                message: Not found
                data: null
                errors: null
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                status: error
                message: Validation error
                data: null
                errors:
                  email:
                    - The email field is required.
  schemas:
    SuccessResponse:
      type: object
      required: [status, message, data, errors]
      properties:
        status:
          type: string
          enum: [success]
          example: success
        message:
          type: string
          example: OK
        data:
          nullable: true
          oneOf:
            - type: object
            - type: array
              items: {}
          example: null
        errors:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/ValidationErrors'
            - type: object
          example: null
    ErrorResponse:
      type: object
      required: [status, message, data, errors]
      properties:
        status:
          type: string
          enum: [error]
          example: error
        message:
          type: string
          example: Validation error
        data:
          nullable: true
          oneOf:
            - type: object
            - type: array
              items: {}
          example: null
        errors:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/ValidationErrors'
            - type: object
          example:
            email:
              - The email field is required.

    # Backward-compatible aliases (used in existing @OA annotations)
...
    ValidationErrors:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: student@example.com
        password:
          type: string
          format: password
          example: password123
    LoginResponseData:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          example: 1|y0ur_sanctum_t0ken
        user:
          $ref: '#/components/schemas/User'
...
    UpdateMeRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        telephone_number:
          type: string
          example: '081234567890'
        gender:
          type: string
          enum: [male, female]
          example: male
        date_of_birth:
          type: string
          format: date
          example: 2005-01-15
        profile_photo:
          type: string
          format: binary
...
    Review:
      type: object
      required: [id, tutor_id, student_id, rate, comment, created_at]
      properties:
        id:
          type: integer
          example: 101
        tutor_id:
          type: integer
          example: 12
        student_id:
          type: integer
          example: 55
        rate:
          type: number
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Tutor sangat membantu dan jelas.
        created_at:
          type: string
          format: date-time
          example: '2026-04-17T10:30:00Z'
    ReviewUpsertRequest:
      type: object
      required: [tutor_id, schedule_id, rate, comment]
      properties:
        tutor_id:
          type: integer
          example: 12
        schedule_id:
          type: integer
          description: ID sesi/jadwal yang direview
          example: 900
        rate:
          type: number
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          example: Penjelasan mudah dipahami.
...
    User:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          enum: [admin, student, tutor, parent, undefined]
          example: student
        telephone_number:
          type: string
          nullable: true
          example: '081234567890'
        profile_photo_path:
          type: string
          nullable: true
          example: uploads/photo.jpg
        gender:
          type: string
          nullable: true
          enum: [male, female]
          example: male
        date_of_birth:
          type: string
          nullable: true
          format: date
          example: 2005-01-15
    Student:
      type: object
      required: [user_id, name, gender, address]
      properties:
        user_id:
          type: integer
          example: 55
        name:
          type: string
          example: Siti Aisyah
        gender:
          type: string
          enum: [male, female]
          example: female
        profile_photo_path:
          type: string
          nullable: true
          example: uploads/student.jpg
        address:
          type: string
          example: Bandung, Jawa Barat
    Tutor:
      type: object
      required: [user_id, name, gender, address, tutor]
      properties:
        user_id:
          type: integer
          example: 12
        name:
          type: string
          example: Ahmad Fauzi
        gender:
          type: string
          enum: [male, female]
          example: male
        profile_photo_path:
          type: string
          nullable: true
          example: uploads/tutor.jpg
        address:
          type: string
          example: Jakarta Selatan, DKI Jakarta
        tutor:
          $ref: '#/components/schemas/TutorProfile'
    TutorProfile:
      type: object
      required: [description, education, teaching_mode, status]
      properties:
        description:
          type: string
          example: Berpengalaman mengajar matematika SMP/SMA.
        education:
          type: string
          example: S1 Pendidikan Matematika
        teaching_mode:
          type: string
          enum: [online, offline]
          example: offline
        status:
          type: string
          enum: [pending, verified, rejected]
          example: verified
...
    MeStudentData:
      type: object
      required: [student]
      properties:
        student:
          $ref: '#/components/schemas/Student'
    MeTutorData:
      type: object
      required: [tutor]
      properties:
        tutor:
          $ref: '#/components/schemas/Tutor'